requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.2
GitPython==3.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.5
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
Send2Trash==1.5.0
six==1.12.0
smmap==3.0.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
virtualenv==20.0.15
virtualenv-clone==0.5.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
psycopg2-binary==2.8.5

requests==2.23.0
stem==1.8.0
PyDriller==1.15.2
peewee==3.13.3
lxml==4.5.1